one/src/main/java/com/example/one/service/impl/SellerServiceImpl.java
package com.example.one.service.impl;

import com.example.one.domain.SellerInfo;
import com.example.one.repository.SellerInfoRepository;
import com.example.one.service.SellerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * Created by QianYunlong on 19
 */
@Service
public class SellerServiceImpl implements SellerService {

    @Autowired
    private SellerInfoRepository sellerInfoRepository;

    @Override
    public SellerInfo findSellerInfoByOpenid(String openid) {
        return sellerInfoRepository.findByOpenid(openid);
    }
}
